+++ app/controllers/web/Suggestions.php
@@ -1,0 +1,10 @@
+<?php 
+
+namespace app\controllers\web;
+
+use app\controllers\api\Suggestions as SuggestionsAPIController;
+
+class Suggestions extends SuggestionsAPIController
+{
+	
+}